# docs/swagger.yaml
definitions:
  APIError:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  ConnectResp:
    properties:
      status:
        type: string
    type: object
...
  PairPhoneReq:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  PairPhoneResp:
    properties:
      pairingCode:
        type: string
    type: object
...
  QRResp:
    properties:
      image:
        type: string
      qrCode:
        type: string
    type: object
...
  SessionCreateReq:
    properties:
      accessToken:
        type: string
      appSecret:
        type: string
      businessAccountId:
        type: string
      engine:
        enum:
        - whatsmeow
        - cloud_api
        type: string
      name:
        type: string
      phoneNumberId:
        type: string
      proxy:
        $ref: '#/definitions/SessionProxy'
      settings:
        $ref: '#/definitions/SessionSettings'
      token:
        type: string
      webhook:
        $ref: '#/definitions/WebhookCreateInline'
      webhookVerifyToken:
        type: string
    required:
    - name
    type: object
  SessionCreatedResp:
    properties:
      businessAccountId:
        type: string
      connected:
        type: integer
      createdAt:
        type: string
      engine:
        type: string
      id:
        type: string
      jid:
        type: string
      name:
        type: string
      phoneNumberId:
        type: string
      proxy:
        $ref: '#/definitions/SessionProxy'
      settings:
        $ref: '#/definitions/SessionSettings'
      status:
        type: string
      token:
        type: string
      updatedAt:
        type: string
      webhook:
        $ref: '#/definitions/WebhookResp'
    type: object
  SessionProfileResp:
    properties:
      businessName:
        type: string
      pictureUrl:
        type: string
      platform:
        type: string
      pushName:
        type: string
      status:
        type: string
    type: object
  SessionProxy:
    properties:
      host:
        type: string
      password:
        type: string
      port:
        type: integer
      protocol:
        type: string
      username:
        type: string
    type: object
  SessionResp:
    properties:
      businessAccountId:
        type: string
      businessName:
        type: string
      chatwootEnabled:
        type: boolean
      connected:
        type: integer
      createdAt:
        type: string
      engine:
        type: string
      id:
        type: string
      jid:
        type: string
      name:
        type: string
      phoneNumberId:
        type: string
      platform:
        type: string
      proxy:
        $ref: '#/definitions/SessionProxy'
      pushName:
        type: string
      settings:
        $ref: '#/definitions/SessionSettings'
      status:
        type: string
      updatedAt:
        type: string
    type: object
  SessionSettings:
    properties:
      alwaysOnline:
        type: boolean
      ignoreGroups:
        type: boolean
      ignoreStatus:
        type: boolean
      msgRejectCall:
        type: string
      readMessages:
        type: boolean
      rejectCall:
        type: boolean
    type: object
  SessionStatusResp:
    properties:
      connected:
        type: boolean
      id:
        type: string
      jid:
        type: string
      loggedIn:
        type: boolean
      name:
        type: string
      status:
        type: string
    type: object
  SessionUpdateReq:
    properties:
      accessToken:
        type: string
      appSecret:
        type: string
      businessAccountId:
        type: string
      engine:
        type: string
      name:
        type: string
      phoneNumberId:
        type: string
      proxy:
        $ref: '#/definitions/SessionProxy'
      settings:
        $ref: '#/definitions/SessionSettings'
      webhookVerifyToken:
        type: string
    type: object
...
  WebhookCreateInline:
    properties:
      events:
        items:
          type: string
        type: array
      url:
        type: string
    type: object
  WebhookResp:
    properties:
      createdAt:
        type: string
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      id:
        type: string
      natsEnabled:
        type: boolean
      secret:
        type: string
      sessionId:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
info:
  contact:
    name: wzap Support
    url: https://github.com/felipyfgs/wzap
  description: WhatsApp Multi-Session API powered by whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wzap API
  version: "1.0"
paths:
...
  /sessions:
    get:
      description: Returns all sessions. Token is never included in list responses.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/SessionResp'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: List sessions (Admin Only)
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Creates a new session with an auto-generated or custom token. Returns
        the full session object including the token.
      parameters:
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SessionCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/SessionCreatedResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Create a new session (Admin Only)
      tags:
      - Sessions
  /sessions/{sessionId}:
    delete:
      description: Disconnects and permanently deletes the session and its device
        from the store
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Delete session
      tags:
      - Sessions
    get:
      description: Returns the session identified by :sessionId (name or id). Token
        is not included.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/SessionResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get session
      tags:
      - Sessions
    put:
      consumes:
      - application/json
      description: Updates session fields (name, proxy, settings). Only provided fields
        are changed.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Session update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SessionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/SessionResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Update session
      tags:
      - Sessions
...
  /sessions/{sessionId}/connect:
    post:
      description: Connects a WhatsApp session. Returns status CONNECTED, PAIRING
        (QR required), or CONNECTING.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/ConnectResp'
              type: object
        "409":
          description: QR pairing already pending
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Connect session
      tags:
      - Sessions
...
  /sessions/{sessionId}/disconnect:
    post:
      description: Disconnects the active WhatsApp session without removing the device
        (can reconnect later)
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Disconnect session
      tags:
      - Sessions
...
  /sessions/{sessionId}/logout:
    post:
      description: Logs out the WhatsApp device (requires re-scan), but keeps the
        session record
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Logout session
      tags:
      - Sessions
...
  /sessions/{sessionId}/pair:
    post:
      consumes:
      - application/json
      description: Generates a pairing code for linking without QR scan. Call /connect
        first.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone number
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/PairPhoneReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/PairPhoneResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Pair via phone number
      tags:
      - Sessions
  /sessions/{sessionId}/profile:
    get:
      description: Returns the WhatsApp profile info for this session (push name,
        picture, bio). Requires an active connection.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/SessionProfileResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get WhatsApp profile
      tags:
      - Sessions
...
  /sessions/{sessionId}/qr:
    get:
      description: Returns the current QR code string and a base64 PNG image. Call
        /connect first, then poll this endpoint until a code is available.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/QRResp'
              type: object
        "404":
          description: No QR code available yet
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get QR code for pairing
      tags:
      - Sessions
  /sessions/{sessionId}/reconnect:
    post:
      description: Disconnects and reconnects the session
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Reconnect session
      tags:
      - Sessions
  /sessions/{sessionId}/restart:
    post:
      description: Disconnects and reconnects the session without losing state
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/SessionResp'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Restart session
      tags:
      - Sessions
  /sessions/{sessionId}/status:
    get:
      description: Returns detailed connection status (connected, loggedIn, JID)
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/SessionStatusResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get session status
      tags:
      - Sessions
...
securityDefinitions:
  Authorization:
    description: Session or admin token (without Bearer prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"